Semantic Engine Website (#35)
* Adjusted header and add landing page layout

* Added accordions

* Added quick links on right handside

* Modified navbar for the new design

* Adjusted footer and Introduction section styling

* Added some navigations

* Added logic for droping a zip file in the landing page

* Fix redirect pages bugs

* Refactored allowedPages in Home page

* Added file type constraints to the drop zone

* Removed commented codes
