+++ src/Footer/Footer.js
@@ -4,32 +4,38 @@ import logoUoG from '../assets/uog-logo.png';
 import logoAgri from '../assets/agri-logo.png';
 import logoSE from '../assets/se-logo.png';
 
-export default function Footer() {
+export default function Footer({ currentPage }) {
   return (
     <>
-    <Divider orientation="horizontal" flexItem />
-    <Stack
-      direction="row"
-      justifyContent="end"
-      sx={{
-        width: "100%",
-      }}
-    >
+      <Divider orientation="horizontal" flexItem />
       <Stack
         direction="row"
+        justifyContent="end"
         sx={{
-          flexDirection: "row",
-          gap: "2rem",
-          justifyContent: "space-between",
-          padding: "2rem",
-          cursor: 'pointer'
+          width: "100%",
         }}
       >
-        <img src={logoUoG} style={{ height: '60px' }} alt="University of Guelph Logo" onClick={() => window.open('https://www.uoguelph.ca/', '_blank')} />
-        <img src={logoAgri} style={{ height: '60px' }} alt="Agri Logo" onClick={() => window.open('https://agrifooddatacanada.ca/', '_blank')} />
-        <img src={logoSE} style={{ height: '60px' }} alt="Semantic Engine Logo" onClick={() => window.open('https://www.semanticengine.org/#/', '_blank')} />
-      </Stack>
-    </Stack >
+        <Stack
+          direction="row"
+          sx={{
+            flexDirection: "row",
+            gap: "2rem",
+            justifyContent: "space-between",
+            padding: "2rem",
+            cursor: 'pointer'
+          }}
+        >
+          {currentPage === 'Landing' ?
+            <>
+              <img src={logoAgri} style={{ height: '80px' }} alt="Agri Logo" onClick={() => window.open('https://agrifooddatacanada.ca/', '_blank')} />
+            </>
+            : <>
+              <img src={logoUoG} style={{ height: '60px' }} alt="University of Guelph Logo" onClick={() => window.open('https://www.uoguelph.ca/', '_blank')} />
+              <img src={logoAgri} style={{ height: '60px' }} alt="Agri Logo" onClick={() => window.open('https://agrifooddatacanada.ca/', '_blank')} />
+              <img src={logoSE} style={{ height: '60px' }} alt="Semantic Engine Logo" onClick={() => window.open('https://www.semanticengine.org/#/', '_blank')} />
+            </>}
+        </Stack>
+      </Stack >
     </>
   );
 };
