+++ src/Header/Header.js
@@ -1,14 +1,16 @@
 import React, { useState, useEffect } from "react";
-import { Typography, Stack, Tooltip, Button, Box } from "@mui/material";
+import { Typography, Tooltip, Button, Box } from "@mui/material";
 import { CustomPalette } from "../constants/customPalette";
 import HelpOutlineIcon from "@mui/icons-material/HelpOutline";
 import logo from '../assets/agri-logo.png';
+import logoWhite from '../assets/agri-logo-white.png';
 import { useLocation } from 'react-router-dom';
+import HeaderWrapper from "./HeaderWrapper";
 
 export default function Header({ currentPage }) {
   const [header, setHeader] = useState(currentPage);
   const [toolTipText, setToolTipText] = useState("");
-  const [helpLink, setHelpLink] = useState("start_schema_help");
+  const [helpLink, setHelpLink] = useState("");
   const location = useLocation();
 
   //Sets headers and tooltip Text based on current page
@@ -82,79 +84,86 @@ export default function Header({ currentPage }) {
         setHelpLink('/overlays_help');
         break;
       default:
-        setHeader("Start Schema");
-        setHelpLink('/start_schema_help');
+        setHeader("");
+        setHelpLink('');
     }
   }, [currentPage]);
 
   return (
-    <Stack
-      direction="row"
-      justifyContent="space-between"
-      sx={{
-        mb: 2,
-        pl: 4,
-        pr: 4,
-        borderBottom: 0.5,
-        borderColor: CustomPalette.GREY_300,
-      }}
-    >
-
-      <Box sx={{ display: "flex", alignItems: "center" }}>
-        <img src={logo} style={{ width: '150px', marginRight: '20px' }} alt="Logo" />
-        <Typography
-          sx={{
-            fontSize: 25,
-            fontWeight: "bold",
-            color: CustomPalette.PRIMARY,
-            alignSelf: "center",
-          }}
-        >
-          {header}
-        </Typography>
-        {toolTipText.length > 0 && (
-          <Box sx={{ marginLeft: 2, color: CustomPalette.GREY_600 }}>
-            <Tooltip
-              title={toolTipText}
-              placement={
-                header === "Attribute Details" || header === "View Schema"
-                  ? "right"
-                  : "right-start"
-              }
-              arrow
+    <HeaderWrapper
+      headerColor={currentPage === "Landing" && CustomPalette.PRIMARY}
+      leftItem={
+        <>
+          {currentPage === "Landing" ?
+            <Typography
+              sx={{
+                fontSize: 40,
+                fontWeight: "bold",
+                color: 'white',
+                alignSelf: "center",
+              }}
             >
-              <HelpOutlineIcon sx={{ fontSize: 15 }} />
-            </Tooltip>
-          </Box>
-        )}
-      </Box>
-      <Stack
-        direction="row"
-        sx={{
-          width: 250,
-          alignItems: "center",
-        }}
-      >
-        {!location.pathname.includes('_help') && (
-          <Button
-            color='button'
-            variant='contained'
-            target='_blank'
-            sx={{
-              m: 2,
-              mr: 5,
-              p: 1,
-              width: '15rem',
-            }}
-            onClick={() =>
-              window.open(`${helpLink}`, '_blank', 'rel=noopener noreferrer')
-            }
-          >
-            Help with this page
-          </Button>
-        )}
-        <Box sx={{ color: CustomPalette.PRIMARY }}>En</Box>
-      </Stack>
-    </Stack >
+              Semantic Engine
+            </Typography>
+            : <>
+              <img src={logo} style={{ width: '150px', marginRight: '20px' }} alt="Logo" />
+              <Typography
+                sx={{
+                  fontSize: 25,
+                  fontWeight: "bold",
+                  color: CustomPalette.PRIMARY,
+                  alignSelf: "center",
+                }}
+              >
+                {header}
+              </Typography>
+              {toolTipText.length > 0 && (
+                <Box sx={{ marginLeft: 2, color: CustomPalette.GREY_600 }}>
+                  <Tooltip
+                    title={toolTipText}
+                    placement={
+                      header === "Attribute Details" || header === "View Schema"
+                        ? "right"
+                        : "right-start"
+                    }
+                    arrow
+                  >
+                    <HelpOutlineIcon sx={{ fontSize: 15 }} />
+                  </Tooltip>
+                </Box>
+              )}
+            </>}
+        </>
+      }
+      rightItem={
+        <>
+          {currentPage === "Landing" ?
+            <img src={logoWhite} style={{ width: '250px', marginRight: '20px' }} alt="Logo" />
+            :
+            <>
+              {!location.pathname.includes('_help') && helpLink !== "" && (
+                <Button
+                  color='button'
+                  variant='contained'
+                  target='_blank'
+                  sx={{
+                    m: 2,
+                    mr: 5,
+                    p: 1,
+                    width: '15rem',
+                  }}
+                  onClick={() =>
+                    window.open(`${helpLink}`, '_blank', 'rel=noopener noreferrer')
+                  }
+                >
+                  Help with this page
+                </Button>
+              )}
+              <Box sx={{ color: CustomPalette.PRIMARY }}>En</Box>
+            </>
+          }
+        </>
+      }
+    />
   );
 }
