+++ src/Home.js
@@ -10,8 +10,18 @@ import CreateManually from "./CreateManually/CreateManually";
 import Overlays from "./Overlays/Overlays";
 import CharacterEncoding from "./Overlays/CharacterEncoding";
 import RequiredEntries from "./Overlays/RequiredEntries";
+import { pagesArray } from "./App";
+import { useMemo } from "react";
 
-const Home = ({ currentPage, pageForward, pageBack, showIntroCard, setShowIntroCard }) => {
+const Home = ({ currentPage, setCurrentPage, pageForward, pageBack, showIntroCard, setShowIntroCard }) => {
+  const allowedPages = useMemo(() => {
+    return [...pagesArray, "Codes", "Create", "Overlays", "CharacterEncoding", "RequiredEntries"];
+  }, []);
+
+  if (!allowedPages.includes(currentPage)) {
+    setCurrentPage("Start");
+  }
+
   return (
     <Box sx={{ flex: 1 }}>
       {currentPage === "Start" && (
